commits changes` I wanted to reach out and ask for some more information. At this time I am unaware of any planned  End of the year Personnel reviews. Which in the past have been completed by Mid January. Have I missed any communication about those reviews?
These reviews are important to me because as per typical operations this is when Merit increases and promotions to new positions are discussed. I am in the middle of a goal to grow myself into a mid-level position and any update on the progress of that goal would be helpful.

I also am growing somewhat concerned about my status on “the bench”. As I have been off of a project for over a month now I have not had the opportunity to try to join any other projects. I do not know what is a typical time to be on the bench, or how long I can continue to remain on the bench. Are there any standards for bench time that can give me some context on what is typical for AT bench time?